feat: changed and updated the models
reformatted the models and added '__str__' and 'class meta' to them in order to display verbose names properly in the admin page.  Created a method inside the classes in order to render the class name in the template view.  Did a simple query and passed the object to render in the template to make sure everything is working properlu.